+++ src/renderer/components/ResponsivePanels/EnvironmentPerformanceGraph.tsx
