+++ src/renderer/components/ResponsivePanels/EnvironmentServerInfo.tsx
